[Chore/#175] Presentation 폴더링 개편 - View, ViewModel
