+++ Staccato-iOS/Staccato-iOS/Model/Error/StaccatoError.swift
