+++ Staccato-iOS/Staccato-iOS/Presentation/Common/Component/CameraView.swift
